blueprints/phase1/requirements.txt
# Vibe Photos Phase 1 — dependency snapshot
# Last updated: 2024-11
# Python version requirement: ==3.12.*
# Mirrors the root `pyproject.toml` + `uv.lock`; keep versions aligned.

# ==================== Core dependencies ====================
fastapi==0.121.1           # Web API framework
uvicorn==0.38.0            # ASGI server
streamlit==1.51.0          # Streamlit UI framework
sqlalchemy==2.0.44         # Database ORM
pillow==11.3.0             # Image manipulation toolkit
python-multipart==0.0.20   # Multipart upload support (aligned with lock file)
aiofiles==24.1.0           # Async file IO
pydantic==2.11.10          # Data validation and serialization
requests==2.32.3           # HTTP client for model downloads
pyyaml==6.0.2              # YAML configuration parser

# ==================== Perception stack ====================
# Primary solution: SigLIP + BLIP (~85% accuracy, multilingual)
torch==2.9.1               # PyTorch deep learning runtime
torchvision==0.24.1        # PyTorch computer vision helpers
transformers==4.57.1       # Hugging Face transformers for SigLIP/BLIP
sentence-transformers==5.1.2  # Sentence embeddings for semantic search prep
# Models download automatically:
# - google/siglip-base-patch16-224-i18n (~400MB)
# - Salesforce/blip-image-captioning-base (~990MB)

# ==================== OCR stack ====================
paddlepaddle==3.2.1        # PaddlePaddle runtime
paddleocr==3.3.1           # PaddleOCR text recognition

# ==================== Database and caching ====================
# SQLite ships with Python; no extra package needed
numpy==2.3.4              # Numerical backbone (transformers dependency)

# ==================== Tooling ====================
# See requirements-dev.txt
